+++ gemrb/plugins/Core/TileMap.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/TileMap.h,v 1.28 2005/02/27 19:13:24 edheldil Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/TileMap.h,v 1.29 2005/03/05 10:03:49 guidoj Exp $
  *
  */
 
@@ -54,7 +54,7 @@ private:
 public:
 	TileMap(void);
 	~TileMap(void);
-	Door* AddDoor(char* Name, unsigned long Flags, int ClosedIndex,
+	Door* AddDoor(char* Name, unsigned int Flags, int ClosedIndex,
 		unsigned short* indexes, int count, Gem_Polygon* open,
 		Gem_Polygon* closed);
 	Door* GetDoor(Point &position);
